requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
djangorestframework==3.12.4
drf-spectacular==0.20.1
inflection==0.5.1
jsonschema==4.0.1
pyrsistent==0.18.0
pytz==2021.3
PyYAML==5.4.1
sqlparse==0.4.2
uritemplate==3.0.1
